fix: apply nullaway
