+++ src/main/java/io/github/jbellis/brokk/gui/Chrome.java
@@ -1531,9 +1531,7 @@ public class Chrome implements AutoCloseable, IConsoleIO, IContextManager.Contex
      */
     @Override
     public void disableHistoryPanel() {
-        if (historyOutputPanel != null) {
-            historyOutputPanel.disableHistory();
-        }
+        historyOutputPanel.disableHistory();
     }
 
     /**
@@ -1553,9 +1551,7 @@ public class Chrome implements AutoCloseable, IConsoleIO, IContextManager.Contex
     public void blockLlmOutput(boolean blocked) {
         // Ensure that prev setText calls are processed before blocking => we need the invokeLater
         SwingUtilities.invokeLater(() -> {
-            if (historyOutputPanel != null) {
-                historyOutputPanel.setMarkdownOutputPanelBlocking(blocked);
-            }
+            historyOutputPanel.setMarkdownOutputPanelBlocking(blocked);
         });
     }
