+++ src/test/java/io/github/jbellis/brokk/gui/mop/stream/SymbolBadgeCustomizerTest.java
@@ -1,8 +1,5 @@
 package io.github.jbellis.brokk.gui.mop.stream;
 
-import io.github.jbellis.brokk.AnalyzerWrapper;
-import io.github.jbellis.brokk.IContextManager;
-import io.github.jbellis.brokk.testutil.MockAnalyzerWrapper;
 import org.jsoup.Jsoup;
 import org.jsoup.nodes.Document;
 import org.jsoup.nodes.Element;
@@ -15,18 +12,6 @@ import static org.junit.jupiter.api.Assertions.*;
  * Simple sanity-checks for {@link SymbolBadgeCustomizer}.
  */
 public class SymbolBadgeCustomizerTest {
-
-    /**
-     * Mock ContextManager that provides a ready analyzer for testing
-     */
-    public static class MockContextManager implements IContextManager {
-        private final MockAnalyzerWrapper analyzerWrapper = new MockAnalyzerWrapper();
-
-        @Override
-        public AnalyzerWrapper getAnalyzerWrapper() {
-            return analyzerWrapper;
-        }
-    }
 
     private Element apply(String html) {
         Document doc = Jsoup.parse(html);
@@ -61,25 +46,4 @@ public class SymbolBadgeCustomizerTest {
                      "Inline code symbol should become a clickable symbol badge");
     }
     
-    @Test
-    public void testFactoryReturnsDefaultWhenAnalyzerWrapperIsNull() {
-        // Create a mock context manager that returns null for analyzer wrapper
-        var mockContextManager = new IContextManager() {
-            @Override
-            public AnalyzerWrapper getAnalyzerWrapper() {
-                return null; // Simulate uninitialized state
-            }
-        };
-        
-        var customizer = SymbolBadgeCustomizer.create(mockContextManager);
-        assertEquals(HtmlCustomizer.DEFAULT, customizer, 
-                     "Should return DEFAULT customizer when analyzer wrapper is null");
-    }
-    
-    @Test
-    public void testFactoryReturnsDefaultWhenContextManagerIsNull() {
-        var customizer = SymbolBadgeCustomizer.create(null);
-        assertEquals(HtmlCustomizer.DEFAULT, customizer, 
-                     "Should return DEFAULT customizer when context manager is null");
-    }
 }
