+++ src/test/java/io/github/jbellis/brokk/gui/search/MarkdownSearchableComponentBasicTest.java
@@ -1,6 +1,7 @@
 package io.github.jbellis.brokk.gui.search;
 
 import io.github.jbellis.brokk.gui.mop.MarkdownOutputPanel;
+import io.github.jbellis.brokk.gui.mop.stream.MockContextManager;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.parallel.Execution;
@@ -33,7 +34,7 @@ public class MarkdownSearchableComponentBasicTest {
 
     @Test
     void testEmptyPanelListHandling() {
-        var emptySearchComponent = new MarkdownSearchableComponent(List.of(), null);
+        var emptySearchComponent = new MarkdownSearchableComponent(List.of(), new MockContextManager());
 
         // Should handle empty panel list gracefully
         assertDoesNotThrow(() -> emptySearchComponent.highlightAll("test", false));
@@ -46,7 +47,7 @@ public class MarkdownSearchableComponentBasicTest {
 
     @Test
     void testCallbackNotification() throws Exception {
-        searchComponent = new MarkdownSearchableComponent(List.of(panel1), null);
+        searchComponent = new MarkdownSearchableComponent(List.of(panel1), new MockContextManager());
 
         CountDownLatch searchComplete = new CountDownLatch(1);
         AtomicBoolean callbackCalled = new AtomicBoolean(false);
@@ -70,5 +71,50 @@ public class MarkdownSearchableComponentBasicTest {
         assertTrue(callbackCalled.get(), "Callback should be called");
         assertEquals(0, totalMatches.get(), "Should have 0 matches for empty panels");
         assertEquals(0, currentMatch.get(), "Should have current match index 0 for no matches");
+    }
+
+    @Test
+    void testNullAndSpecialCharacterSearch() throws Exception {
+        searchComponent = new MarkdownSearchableComponent(List.of(panel1), new MockContextManager());
+
+        // Test special characters
+        String[] specialSearches = {"$", "@#", "^&*", "()", "\\n", "\t"};
+
+        for (String special : specialSearches) {
+            final String searchTerm = special;
+            CountDownLatch searchDone = new CountDownLatch(1);
+            searchComponent.setSearchCompleteCallback((total, current) -> {
+                searchDone.countDown();
+            });
+
+            SwingUtilities.invokeLater(() ->
+                                               searchComponent.highlightAll(searchTerm, false));
+
+            assertTrue(searchDone.await(10, TimeUnit.SECONDS),
+                       "Search for '" + searchTerm + "' should complete");
+        }
+    }
+
+    @Test
+    void testSearchWithNoMatches() throws Exception {
+        searchComponent = new MarkdownSearchableComponent(List.of(panel1), new MockContextManager());
+
+        CountDownLatch searchComplete = new CountDownLatch(1);
+        AtomicInteger totalMatches = new AtomicInteger(-1);
+
+        searchComponent.setSearchCompleteCallback((total, current) -> {
+            totalMatches.set(total);
+            searchComplete.countDown();
+        });
+
+        // Search for non-existent term
+        SwingUtilities.invokeLater(() -> searchComponent.highlightAll("nonexistentterm123", false));
+
+        assertTrue(searchComplete.await(10, TimeUnit.SECONDS));
+        assertEquals(0, totalMatches.get(), "Should have no matches");
+
+        // Navigation should fail
+        assertFalse(searchComponent.findNext("nonexistentterm123", false, true),
+                    "Navigation should fail with no matches");
     }
 }
